alertmanager: Add Telegram Receiver type to AlertmanagerConfig CRD
Changes chatID from secret to int and adds a check for the alertmanager version

Fixes: #4703

+++ pkg/alertmanager/types.go
@@ -322,7 +322,7 @@ type telegramConfig struct {
 	VSendResolved        *bool             `yaml:"send_resolved,omitempty" json:"send_resolved,omitempty"`
 	APIUrl               string            `yaml:"api_url,omitempty" json:"api_url,omitempty"`
 	BotToken             string            `yaml:"bot_token,omitempty" json:"bot_token,omitempty"`
-	ChatID               int32             `yaml:"chat_id,omitempty" json:"chat_id,omitempty"`
+	ChatID               int64             `yaml:"chat_id,omitempty" json:"chat_id,omitempty"`
 	Message              string            `yaml:"message,omitempty" json:"message,omitempty"`
 	DisableNotifications bool              `yaml:"disable_notifications,omitempty" json:"disable_notifications,omitempty"`
 	ParseMode            string            `yaml:"parse_mode,omitempty" json:"parse_mode,omitempty"`
